requirements.txt
kaggle
python-dotenv
pandas
numpy
seaborn
matplotlib
scikit-learn
joblib
streamlit
